+++ packages/admin-ui/ui/src/domain/products/edit/sections/variants/table.tsx
@@ -106,9 +106,9 @@ const VariantsTable = ({ variants, actions }: Props) => {
 
   const getTableRowActionables = (variant: ProductVariant) => {
     const inventoryManagementActions = []
-    if (hasInventoryService && variant.manage_inventory) {
+    if (hasInventoryService) {
       inventoryManagementActions.push({
-        label: "Manage inventory", // TODO: Only add this item if variant.manageInventory is true
+        label: "Manage inventory",
         icon: <BuildingsIcon size="20" />,
         onClick: () => updateVariantInventory(variant),
       })