+++ packages/inventory/src/index.ts
@@ -1,12 +1,14 @@
-import ConnectionLoader from "./loaders/connection"
-import InventoryService from "./services/inventory"
+import loadConnection from "./loaders/connection"
+import loadContainer from "./loaders/container"
+
+import migrations from "./migrations"
 import * as InventoryModels from "./models"
-import * as SchemaMigration from "./migrations/schema-migrations/1665748086258-inventory_setup"
+import InventoryService from "./services/inventory"
+
 import { ModuleExports } from "@medusajs/modules-sdk"
 
 const service = InventoryService
-const migrations = [SchemaMigration]
-const loaders = [ConnectionLoader]
+const loaders = [loadContainer, loadConnection]
 const models = Object.values(InventoryModels)
 
 const moduleDefinition: ModuleExports = {